Refactor supplier management: restructure file organization
