+++ backend/src/suppliers/dtos/create-supplier.dto.ts
@@ -1,8 +1,7 @@
 import { z } from 'zod';
-import { supplierStatusValues } from '../constants';
+import { supplierStatusValues } from '../../constants';
 
-export const supplierSchema = z.object({
-  id: z.string(),
+export const createSupplierSchema = z.object({
   status: z.enum(supplierStatusValues),
   companyName: z
     .string()
@@ -17,8 +16,6 @@ export const supplierSchema = z.object({
   phoneNumber: z.string().optional(),
   productsProvided: z.array(z.string()),
   branches: z.array(z.string()),
-  createdAt: z.string(),
-  updatedAt: z.string(),
 });
 
-export type SupplierDto = z.infer<typeof supplierSchema>;
+export type CreateSupplierDto = z.infer<typeof createSupplierSchema>;
