+++ backend/src/utils/mappers/supplier.mapper.ts
@@ -1,5 +1,5 @@
-import { ISupplier } from '../../models/supplier.model';
-import { SupplierDto } from '../../dtos/supplier.dto';
+import { ISupplier } from '../../suppliers/supplier.model';
+import { SupplierDto } from '../../suppliers/dtos/supplier.dto';
 
 export const toSupplierResponseDTO = (supplier: ISupplier): SupplierDto => ({
   id: supplier._id.toString(),
@@ -8,7 +8,7 @@ export const toSupplierResponseDTO = (supplier: ISupplier): SupplierDto => ({
   address: supplier.address ?? '',
   email: supplier.email,
   phoneNumber: supplier.phoneNumber ?? '',
-  status: supplier.status as "Active" | "Inactive" | "Pending",
+  status: supplier.status as 'Active' | 'Inactive' | 'Pending',
   productsProvided: supplier.productsProvided.map((product) =>
     product.toString(),
   ),
